+++ main.go
@@ -319,7 +319,7 @@ func main() {
 
 		targetGroupConfigurationReconciler := gateway.NewTargetGroupConfigurationReconciler(
 			mgr.GetClient(),
-			mgr.GetEventRecorderFor(gateway_constants.LoadBalancerConfigurationController),
+			mgr.GetEventRecorderFor(gateway_constants.TargetGroupConfigurationController),
 			controllerCFG,
 			serviceReferenceCounter,
 			finalizerManager,
@@ -335,6 +335,26 @@ func main() {
 		err = targetGroupConfigurationReconciler.SetupWatches(ctx, tgCfgController, mgr)
 		if err != nil {
 			setupLog.Error(err, "Unable to set up TargetGroupConfiguration Watches")
+			os.Exit(1)
+		}
+
+		listenerRuleConfigurationReconciler := gateway.NewListenerRuleConfigurationReconciler(
+			mgr.GetClient(),
+			mgr.GetEventRecorderFor(gateway_constants.ListenerRuleConfigurationController),
+			controllerCFG,
+			finalizerManager,
+			mgr.GetLogger().WithName("listenerruleconfiguration-controller"),
+		)
+
+		listenerRuleCfgController, err := listenerRuleConfigurationReconciler.SetupWithManager(ctx, mgr)
+		if err != nil {
+			setupLog.Error(err, "Unable to set up ListenerRuleConfiguration Manager")
+			os.Exit(1)
+		}
+
+		err = listenerRuleConfigurationReconciler.SetupWatches(ctx, listenerRuleCfgController, mgr)
+		if err != nil {
+			setupLog.Error(err, "Unable to set up ListenerRuleConfiguration Watches")
 			os.Exit(1)
 		}
