+++ pkg/gateway/routeutils/utils.go
@@ -3,12 +3,9 @@ package routeutils
 import (
 	"context"
 	"fmt"
-	"github.com/pkg/errors"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"net"
-	elbv2gw "sigs.k8s.io/aws-load-balancer-controller/apis/gateway/v1beta1"
-	"sigs.k8s.io/aws-load-balancer-controller/pkg/k8s"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"strings"
 )
@@ -33,7 +30,7 @@ func ListL4Routes(ctx context.Context, k8sClient client.Client) ([]preLoadRouteD
 	}
 	l4Routes = append(l4Routes, tlsRoutes...)
 	if len(failedRoutes) > 0 {
-		err = errors.Errorf("failed to list L4 routes, %s", failedRoutes)
+		err = fmt.Errorf("failed to list L4 routes, %v", failedRoutes)
 	}
 	return l4Routes, err
 }
@@ -53,7 +50,7 @@ func ListL7Routes(ctx context.Context, k8sClient client.Client) ([]preLoadRouteD
 	}
 	l7Routes = append(l7Routes, grpcRoutes...)
 	if len(failedRoutes) > 0 {
-		err = errors.Errorf("failed to list L7 routes, %s", failedRoutes)
+		err = fmt.Errorf("failed to list L7 routes, %v", failedRoutes)
 	}
 	return l7Routes, err
 }
@@ -77,21 +74,6 @@ func FilterRoutesBySvc(routes []preLoadRouteDescriptor, svc *corev1.Service) []p
 	return filteredRoutes
 }
 
-// FilterRoutesByListenerRuleCfg filters a slice of routes based on ListenerRuleConfiguration reference.
-// Returns a new slice containing only routes that reference the specified ListenerRuleConfiguration.
-func FilterRoutesByListenerRuleCfg(routes []preLoadRouteDescriptor, ruleConfig *elbv2gw.ListenerRuleConfiguration) []preLoadRouteDescriptor {
-	if ruleConfig == nil || len(routes) == 0 {
-		return []preLoadRouteDescriptor{}
-	}
-	filteredRoutes := make([]preLoadRouteDescriptor, 0, len(routes))
-	for _, route := range routes {
-		if isListenerRuleConfigReferredByRoute(route, k8s.NamespacedName(ruleConfig)) {
-			filteredRoutes = append(filteredRoutes, route)
-		}
-	}
-	return filteredRoutes
-}
-
 // isServiceReferredByRoute checks if a route references a specific service.
 // Assuming we are only supporting services as backendRefs on Routes
 func isServiceReferredByRoute(route preLoadRouteDescriptor, svcID types.NamespacedName) bool {
@@ -102,17 +84,6 @@ func isServiceReferredByRoute(route preLoadRouteDescriptor, svcID types.Namespac
 		}
 
 		if string(backendRef.Name) == svcID.Name && namespace == svcID.Namespace {
-			return true
-		}
-	}
-	return false
-}
-
-// isListenerRuleConfigReferredByRoute checks if a route references a specific ruleConfig.
-func isListenerRuleConfigReferredByRoute(route preLoadRouteDescriptor, ruleConfig types.NamespacedName) bool {
-	for _, config := range route.GetListenerRuleConfigs() {
-		namespace := route.GetRouteNamespacedName().Namespace
-		if string(config.Name) == ruleConfig.Name && namespace == ruleConfig.Namespace {
 			return true
 		}
 	}
